pytorch-model-deployment/requirements.txt
flask
flask-restful
pillow
requests
torch
torchvision
